% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bto_codes}
\alias{bto_codes}
\title{This is data to be included in my package}
\description{
The British Trust for Ornithology has a list of 2-letter species codes for over 250 bird species.
The dataset acts as a lookup table, matching species names with their corresponding BTO codes.
}
\references{
\url{ https://www.bto.org/sites/default/files/u16/downloads/forms_instructions/bto_bird_species_codes.pdf}
}
\keyword{data}
