% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_list}
\alias{location_list}
\title{This is data to be included in my package}
\description{
This has a list of recorders that were deployed along with accompanying metadata about the sites studied.
If the current audio file in the above format matches a recorder in the list,
the following columns of the file are passed to the metadata tab:
\itemize{
\item recorder name: prefix for audio file names recorded by this device
\item lat: latitude of the recorder for the study period, in decimal degrees
\item long: longitude of the recorder for the study period, in decimal degrees
\item location name: name of the study site
\item location county: county of study site
\item habitat type: primary habitat type of the study site
\item dist to coastline: approximate distance to the nearest coastline in kilometres
}
If any of this information is unavailable, the column name in the metadata panel will still appear but the
body of text will be blank. Extra columns can be added to the file, where they will be printed verbatim to
the metadata panel.
}
\keyword{data}
