% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_list}
\alias{species_list}
\title{This is data to be included in my package}
\description{
NEAL allows for multiple species lists to be included, all of which are stored in species_list.csv.
The file has one column for each site being studied, with the first entry of each column being the site
name. Columns can have different numbers of rows (species), i.e. all those one expects to find at the
given site. The species names can be stored in the column in any order as they are sorted alphabetically at
run time. Columns can be appended to it by uploading via a widget in the Configuration tab. Each of the
columns in the uploaded data not already present in the species data are appended.
}
\keyword{data}
